backend/requirements/base.txt
Django==2.1.4
djangorestframework==3.9.0
psycopg2==2.7.6.1
marshmallow==2.16.3
aldjemy==0.9.0
PyYAML==3.12